backend/services/service-documents/requirements.txt
quart
motor
python-dotenv
pydantic
pydash
pytest
pytest-asyncio
pytest-mock
azure-storage-blob